chore: remove text resize
