+++ frontend/src/features/admin-form/preview/PreviewFormPage.tsx
@@ -1,4 +1,3 @@
-import { useEffect, useState } from 'react'
 import { useParams } from 'react-router-dom'
 import { Flex } from '@chakra-ui/react'
 
@@ -14,38 +13,14 @@ import { FormFooter } from '~features/public-form/components/FormFooter'
 import FormInstructions from '~features/public-form/components/FormInstructions'
 import { PublicFormLogo } from '~features/public-form/components/FormLogo'
 import FormStartPage from '~features/public-form/components/FormStartPage'
+import LanguageControl from '~features/public-form/components/LanguageControl'
 import { PublicFormWrapper } from '~features/public-form/components/PublicFormWrapper'
-import ZoomControl from '~features/public-form/components/ZoomControl'
 
 import { PreviewFormBannerContainer } from '../common/components/PreviewFormBanner'
 
 import { PreviewFormProvider } from './PreviewFormProvider'
 
 export const PreviewFormPage = (): JSX.Element => {
-  const [rootFontSize, setRootFontSize] = useState('16px')
-
-  const setDefaultSize = () => {
-    setRootFontSize('16px')
-  }
-
-  const setLargeSize = () => {
-    setRootFontSize('18px')
-  }
-
-  const setLargestSize = () => {
-    setRootFontSize('24px')
-  }
-
-  useEffect(() => {
-    const root = document.documentElement
-    root.style.setProperty('font-size', rootFontSize)
-
-    // need to reset on unmount
-    return () => {
-      root.style.removeProperty('font-size')
-    }
-  }, [rootFontSize])
-
   const { formId } = useParams()
   if (!formId) throw new Error('No formId provided')
 
@@ -57,11 +32,7 @@ export const PreviewFormPage = (): JSX.Element => {
         <FormSectionsProvider>
           <PublicFormLogo />
           <FormStartPage />
-          <ZoomControl
-            setDefaultSize={setDefaultSize}
-            setLargeSize={setLargeSize}
-            setLargestSize={setLargestSize}
-          />
+          <LanguageControl />
           <PublicFormWrapper>
             <FormInstructions />
             <FormFields />
