+++ frontend/src/features/public-form/components/LanguageControl/FontLargestSvgr.tsx
