+++ frontend/src/features/public-form/components/LanguageControl/index.ts
@@ -1,0 +1,1 @@
+export { LanguageControl as default } from './LanguageControl'
