--- frontend/src/features/public-form/components/ZoomControl/index.ts
@@ -1,1 +1,0 @@
-export { ZoomControl as default } from './ZoomControl'
